+++ monad-chronicle/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1
+
+- Remove `disclose`. Use `dictate c >> return def` if you need it.
+  No more dependency of `data-default` or `data-default-class`.
+  This change is a consequence of change in `data-default-0.8.0.0`
+
 # 1.0.2
 
 - Support GHC-8.6.5...GHC-9.10.1